Move tests for executor features into separate file
